+++ ai/openai_agent.py
@@ -1,5 +1,5 @@
 import openai
-from openai.error import OpenAIError, Timeout
+from openai import OpenAIError, APITimeoutError
 from config import OPENAI_API_KEY
 
 if not OPENAI_API_KEY:
@@ -7,9 +7,10 @@ if not OPENAI_API_KEY:
 
 openai.api_key = OPENAI_API_KEY
 
+
 def suggest_ticket_response(ticket: dict, context: str = "") -> str:
     prompt = (
-        f"You are a Tier 1 help desk agent for a truck stop. Here is the ticket:\n"
+        "You are a Tier 1 help desk agent for a truck stop. Here is the ticket:\n"
         f"{ticket}\n"
         f"Context: {context}\n"
         "Suggest the best response, including troubleshooting or assignment if possible."
@@ -17,20 +18,11 @@ def suggest_ticket_response(ticket: dict, context: str = "") -> str:
     try:
         response = openai.ChatCompletion.create(
             model="gpt-4o",
-
             messages=[{"role": "system", "content": prompt}],
             timeout=15,
         )
         return response["choices"][0]["message"]["content"]
-    except Timeout:
+    except APITimeoutError:
         return "OpenAI request timed out."
-    except OpenAIError as e:
-        return f"OpenAI API error: {e}"
-
-            messages=[{"role": "system", "content": prompt}]
-        )
-    except openai.error.OpenAIError as exc:
+    except OpenAIError as exc:
         return f"OpenAI API error: {exc}"
-
-    return response['choices'][0]['message']['content']
-
